+++ test/integration/model/geography.test.js
@@ -10,12 +10,12 @@ var chai = require('chai')
 var current = Support.sequelize;
 
 describe(Support.getTestDialectTeaser('Model'), function() {
-  if (current.dialect.supports.GEOMETRY) {
-    describe('GEOMETRY', function() {
+  if (current.dialect.supports.GEOGRAPHY) {
+    describe('GEOGRAPHY', function() {
       beforeEach(function() {
         this.User = this.sequelize.define('User', {
           username: DataTypes.STRING,
-          geometry: DataTypes.GEOMETRY
+          geography: DataTypes.GEOGRAPHY
         });
 
         return this.User.sync({ force: true });
@@ -23,7 +23,7 @@ describe(Support.getTestDialectTeaser('Model'), function() {
 
       it('works with aliases fields', function () {
         var Pub = this.sequelize.define('Pub', {
-          location: {field: 'coordinates', type: DataTypes.GEOMETRY}
+          location: {field: 'coordinates', type: DataTypes.GEOGRAPHY}
         })
           , point = {type: 'Point', coordinates: [39.807222, -76.984722]};
 
@@ -35,128 +35,128 @@ describe(Support.getTestDialectTeaser('Model'), function() {
         });
       });
 
-      it('should create a geometry object', function() {
+      it('should create a geography object', function() {
         var User = this.User;
         var point = { type: 'Point', coordinates: [39.807222,-76.984722]};
 
-        return User.create({username: 'username', geometry: point }).then(function(newUser) {
+        return User.create({username: 'username', geography: point }).then(function(newUser) {
           expect(newUser).not.to.be.null;
-          expect(newUser.geometry).to.be.deep.eql(point);
+          expect(newUser.geography).to.be.deep.eql(point);
         });
       });
 
-      it('should update a geometry object', function() {
+      it('should update a geography object', function() {
         var User = this.User;
         var point1 = { type: 'Point', coordinates: [39.807222,-76.984722]}
           , point2 = { type: 'Point', coordinates: [49.807222,-86.984722]};
-        var props = {username: 'username', geometry: point1};
+        var props = {username: 'username', geography: point1};
 
         return User.create(props).then(function(user) {
-          return User.update({geometry: point2}, {where: {username: props.username}});
+          return User.update({geography: point2}, {where: {username: props.username}});
         }).then(function(count) {
           return User.findOne({where: {username: props.username}});
         }).then(function(user) {
-          expect(user.geometry).to.be.deep.eql(point2);
+          expect(user.geography).to.be.deep.eql(point2);
         });
       });
     });
 
-    describe('GEOMETRY(POINT)', function() {
+    describe('GEOGRAPHY(POINT)', function() {
       beforeEach(function() {
         this.User = this.sequelize.define('User', {
           username: DataTypes.STRING,
-          geometry: DataTypes.GEOMETRY('POINT')
+          geography: DataTypes.GEOGRAPHY('POINT')
         });
 
         return this.User.sync({ force: true });
       });
 
-      it('should create a geometry object', function() {
+      it('should create a geography object', function() {
         var User = this.User;
         var point = { type: 'Point', coordinates: [39.807222,-76.984722]};
 
-        return User.create({username: 'username', geometry: point }).then(function(newUser) {
+        return User.create({username: 'username', geography: point }).then(function(newUser) {
           expect(newUser).not.to.be.null;
-          expect(newUser.geometry).to.be.deep.eql(point);
+          expect(newUser.geography).to.be.deep.eql(point);
         });
       });
 
-      it('should update a geometry object', function() {
+      it('should update a geography object', function() {
         var User = this.User;
         var point1 = { type: 'Point', coordinates: [39.807222,-76.984722]}
           , point2 = { type: 'Point', coordinates: [49.807222,-86.984722]};
-        var props = {username: 'username', geometry: point1};
+        var props = {username: 'username', geography: point1};
 
         return User.create(props).then(function(user) {
-          return User.update({geometry: point2}, {where: {username: props.username}});
+          return User.update({geography: point2}, {where: {username: props.username}});
         }).then(function(count) {
           return User.findOne({where: {username: props.username}});
         }).then(function(user) {
-          expect(user.geometry).to.be.deep.eql(point2);
+          expect(user.geography).to.be.deep.eql(point2);
         });
       });
     });
 
-    describe('GEOMETRY(LINESTRING)', function() {
+    describe('GEOGRAPHY(LINESTRING)', function() {
       beforeEach(function() {
         this.User = this.sequelize.define('User', {
           username: DataTypes.STRING,
-          geometry: DataTypes.GEOMETRY('LINESTRING')
+          geography: DataTypes.GEOGRAPHY('LINESTRING')
         });
 
         return this.User.sync({ force: true });
       });
 
-      it('should create a geometry object', function() {
+      it('should create a geography object', function() {
         var User = this.User;
         var point = { type: 'LineString', 'coordinates': [ [100.0, 0.0], [101.0, 1.0] ] };
 
-        return User.create({username: 'username', geometry: point }).then(function(newUser) {
+        return User.create({username: 'username', geography: point }).then(function(newUser) {
           expect(newUser).not.to.be.null;
-          expect(newUser.geometry).to.be.deep.eql(point);
+          expect(newUser.geography).to.be.deep.eql(point);
         });
       });
 
-      it('should update a geometry object', function() {
+      it('should update a geography object', function() {
         var User = this.User;
         var point1 = { type: 'LineString', coordinates: [ [100.0, 0.0], [101.0, 1.0] ] }
           , point2 = { type: 'LineString', coordinates: [ [101.0, 0.0], [102.0, 1.0] ] };
-        var props = {username: 'username', geometry: point1};
+        var props = {username: 'username', geography: point1};
 
         return User.create(props).then(function(user) {
-          return User.update({geometry: point2}, {where: {username: props.username}});
+          return User.update({geography: point2}, {where: {username: props.username}});
         }).then(function(count) {
           return User.findOne({where: {username: props.username}});
         }).then(function(user) {
-          expect(user.geometry).to.be.deep.eql(point2);
+          expect(user.geography).to.be.deep.eql(point2);
         });
       });
     });
 
-    describe('GEOMETRY(POLYGON)', function() {
+    describe('GEOGRAPHY(POLYGON)', function() {
       beforeEach(function() {
         this.User = this.sequelize.define('User', {
           username: DataTypes.STRING,
-          geometry: DataTypes.GEOMETRY('POLYGON')
+          geography: DataTypes.GEOGRAPHY('POLYGON')
         });
 
         return this.User.sync({ force: true });
       });
 
-      it('should create a geometry object', function() {
+      it('should create a geography object', function() {
         var User = this.User;
         var point = { type: 'Polygon', coordinates: [
              [ [100.0, 0.0], [101.0, 0.0], [101.0, 1.0],
                [100.0, 1.0], [100.0, 0.0] ]
              ]};
 
-        return User.create({username: 'username', geometry: point }).then(function(newUser) {
+        return User.create({username: 'username', geography: point }).then(function(newUser) {
           expect(newUser).not.to.be.null;
-          expect(newUser.geometry).to.be.deep.eql(point);
+          expect(newUser.geography).to.be.deep.eql(point);
         });
       });
 
-      it('should update a geometry object', function() {
+      it('should update a geography object', function() {
         var User = this.User;
         var polygon1 = { type: 'Polygon', coordinates: [
               [ [100.0, 0.0], [101.0, 0.0], [101.0, 1.0],
@@ -166,14 +166,14 @@ describe(Support.getTestDialectTeaser('Model'), function() {
               [ [100.0, 0.0], [102.0, 0.0], [102.0, 1.0],
                 [100.0, 1.0], [100.0, 0.0] ]
               ]};
-        var props = {username: 'username', geometry: polygon1};
+        var props = {username: 'username', geography: polygon1};
 
         return User.create(props).then(function(user) {
-          return User.update({geometry: polygon2}, {where: {username: props.username}});
+          return User.update({geography: polygon2}, {where: {username: props.username}});
         }).then(function(count) {
           return User.findOne({where: {username: props.username}});
         }).then(function(user) {
-          expect(user.geometry).to.be.deep.eql(polygon2);
+          expect(user.geography).to.be.deep.eql(polygon2);
         });
       });
     });